pyproject.toml
[project]
name = "ai-eco-scaffold"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
